requirements.txt
absl-py==2.1.0
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
blinker==1.9.0
cachetools==5.5.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
Django==4.2.3
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
ecdsa==0.18.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0.post1
Faker==33.0.0
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.0
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
graphviz==0.20.1
greenlet==3.1.1
grpcio==1.65.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.4
imageio==2.34.2
inflection==0.5.1
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-huggingface==0.1.2
langchain-ollama==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.143
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.14.1
networkx==3.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
ollama==0.3.3
openai==1.55.0
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.11
packaging==23.2
pandas==2.2.3
phonenumbers==8.13.20
Pillow==10.0.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
protobuf==4.25.4
psycopg2-binary==2.9.8
pyarrow==18.0.0
pyasn1==0.5.0
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pypdf==5.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-keycloak==3.3.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
sqlparse==0.4.4
streamlit==1.40.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.4.0
tornado==6.4.1
tqdm==4.67.0
transformers==4.46.3
triton==3.0.0
twilio==8.8.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.0.4
watchdog==6.0.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.2